[*] Rename folders swapchain and window to plural
