+++ src/vulkan-renderer/wrapper/swapchains/swapchain_utils.cpp
@@ -1,4 +1,4 @@
-#include "inexor/vulkan-renderer/wrapper/swapchain/swapchain_utils.hpp"
+#include "inexor/vulkan-renderer/wrapper/swapchains/swapchain_utils.hpp"
 
 #include "inexor/vulkan-renderer/tools/exception.hpp"
 #include "inexor/vulkan-renderer/tools/representation.hpp"
